refactor: remove api services
