+++ src/pages/main-page/main-page.tsx
@@ -6,8 +6,7 @@ import { useAppDispatch } from '@/hooks/store/useAppDispatch';
 import { useAppSelector } from '@/hooks/store/useAppSelector';
 import useSelectedPoint from '@/hooks/useSelectedPoint';
 import { fetchOffersAction } from '@/store/modules/cities/api-actions';
-import { getPlaces } from '@/store/modules/cities/selectors';
-import { getCurrentCity } from '@/store/modules/common/selectors';
+import { getCurrentCity, getPlaces } from '@/store/modules/cities/selectors';
 
 import Header from '@/components/common/header/header';
 import MapSection from '@/components/common/map-section/map-section';
@@ -15,7 +14,7 @@ import CitiesPlacesEmpty from '~/cities/cities-places-empty/cities-places-empty'
 import CitiesPlaces from '~/cities/cities-places/cities-places';
 import CitiesTabs from '~/cities/cities-tabs/cities-tabs';
 
-import offerApiService from '@/service/offer-api-service';
+import OfferApiService from '@/service/offer-api-service';
 import { MapType } from '@/utils/consts';
 
 function MainPage(): JSX.Element {
@@ -24,7 +23,7 @@ function MainPage(): JSX.Element {
   const dispatch = useAppDispatch();
   const { selectedPointId, handleSelectedPointState } = useSelectedPoint();
 
-  const places = offerApiService.getOffersInCity(offerCards, currentCity);
+  const places = OfferApiService.getOffersInCity(offerCards, currentCity);
   const hasPlaces = !!places.length;
 
   useEffect(() => {
