--- src/service/favorite-api-service.ts
@@ -1,9 +1,0 @@
-import { MOCK_OFFERS_FAVORITES } from '@/mocks/offers';
-
-class FavoriteApiService {
-  get favorite() {
-    return MOCK_OFFERS_FAVORITES;
-  }
-}
-
-export default new FavoriteApiService();
